# public/swaggerApi/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OTC Forentend API",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "{schema}://192.168.10.195:8080",
            "description": "OpenApi parameters",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "http"
                }
            }
        }
    ],
    "paths": {
...
        "/otc/order/create": {
            "post": {
                "tags": [
                    "otc-order"
                ],
                "summary": "创建订单",
                "operationId": "createOrder",
                "parameters": [
                    {
                        "name": "counterId",
                        "in": "query",
                        "description": "广告id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "in",
                        "in": "query",
                        "description": "数量 OR 金额amount=金额（法币）,quantity=数量（数字币）",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "amount",
                                "quantity"
                            ]
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "金额（法币）:in=amount时传递",
                        "required": false,
                        "schema": {
                            "type": "double"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "数量（数字币）:in=quantity时传递",
                        "required": false,
                        "schema": {
                            "type": "double"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "code": 200,
                                    "msg": "成功",
                                    "data": {
                                        "info": {
                                            "id": 7,
                                            "counterId": 4,
                                            "buyerId": 950398,
                                            "sellerId": 950397,
                                            "status": 3,
                                            "price": "24286.83",
                                            "amount": "20000.00",
                                            "quantity": "0.82349158",
                                            "fee": "0.00576444",
                                            "buyerFeedback": 1,
                                            "sellerFeedback": 1,
                                            "expiresAt": "1545974722000",
                                            "createTime": "1545974722000",
                                            "updateTime": "1545974722000"
                                        },
                                        "messages": [
                                            {
                                                "id": 7,
                                                "orderId": 401,
                                                "fromUid": 950398,
                                                "toUid": 950397,
                                                "type": 1,
                                                "isRead": 0,
                                                "content": "hello",
                                                "extra": "",
                                                "createTime": "1545974722000",
                                                "updateTime": "1545974722000"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/order/{orderId}": {
            "get": {
                "tags": [
                    "otc-order"
                ],
                "summary": "订单详情",
                "operationId": "showOrder",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "query",
                        "description": "订单id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "code": 200,
                                    "msg": "成功",
                                    "data": {
                                        "info": {
                                            "id": 7,
                                            "counterId": 4,
                                            "buyerId": 950398,
                                            "sellerId": 950397,
                                            "status": 3,
                                            "price": "24286.83",
                                            "amount": "20000.00",
                                            "quantity": "0.82349158",
                                            "fee": "0.00576444",
                                            "buyerFeedback": 1,
                                            "sellerFeedback": 1,
                                            "expiresAt": "1545974722000",
                                            "createTime": "1545974722000",
                                            "updateTime": "1545974722000"
                                        },
                                        "messages": [
                                            {
                                                "id": 7,
                                                "orderId": 401,
                                                "fromUid": 950398,
                                                "toUid": 950397,
                                                "type": 1,
                                                "isRead": 0,
                                                "content": "hello",
                                                "extra": "",
                                                "createTime": "1545974722000",
                                                "updateTime": "1545974722000"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/order": {
            "get": {
                "tags": [
                    "otc-order"
                ],
                "summary": "订单列表",
                "operationId": "listOrderOTC",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "当前页，默认1",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "当前页大小，默认10",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "排序字段，默认id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "排列顺序，默认desc",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "类型:Processing=进行中的交易,Completed=已完成的交易",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Processing",
                                "Completed"
                            ]
                        }
                    },
                    {
                        "name": "coin",
                        "in": "query",
                        "description": "币种:BTC=比特币,LTC=莱特币",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/order/repeal/{orderId}": {
            "post": {
                "tags": [
                    "otc-order"
                ],
                "summary": "撤销订单",
                "operationId": "repealOrder",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "query",
                        "description": "订单id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/order/restart/{orderId}": {
            "post": {
                "tags": [
                    "otc-order"
                ],
                "summary": "重启订单",
                "operationId": "restartOrder",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "query",
                        "description": "订单id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/order/payment/{orderId}": {
            "post": {
                "tags": [
                    "otc-order"
                ],
                "summary": "已付款",
                "operationId": "paymentOrder",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "query",
                        "description": "订单id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/order/release/{orderId}": {
            "post": {
                "tags": [
                    "otc-order"
                ],
                "summary": "释放托管",
                "operationId": "releaseOrder",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "query",
                        "description": "订单id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "transactionPassword",
                        "in": "query",
                        "description": "交易密码",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/otc/order/review/{orderId}": {
            "post": {
                "tags": [
                    "otc-order"
                ],
                "summary": "评论订单",
                "operationId": "reviewOrder",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "query",
                        "description": "订单id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "feedback",
                        "in": "query",
                        "description": "评价:1=好评,2=中评,3=差评",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "enum": [
                                "1",
                                "2",
                                "3"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
    }
}